third_party/blink/web_tests/fast/frames/sandboxed-iframe-navigation-windowopen-expected.txt

CONSOLE ERROR: Blocked opening 'javascript: opener.parent.windowOpened(opener);' in a new window because the request was made in a sandboxed frame whose 'allow-popups' permission is not set.
Verifies that a sandboxed IFrame cannot open new windows using window.open(). Expect ten windows to be opened, but the sandboxed IFrame not to be the opener of any of them.

This test will print "PASS" on success.

          

PASS
